+++ src/net/l1j/server/clientpackets/C_ChatWhisper.java
@@ -82,8 +82,8 @@ public class C_ChatWhisper extends ClientBasePacket {
 			_log.log(record);
 		}
 
-		whisperFrom.sendPackets(new S_ChatPacket(whisperTo, text, Opcodes.S_OPCODE_GLOBALCHAT, (byte) 9));
-		whisperTo.sendPackets(new S_ChatPacket(whisperFrom, text, Opcodes.S_OPCODE_WHISPERCHAT, (byte) 16));
+		whisperFrom.sendPackets(new S_ChatPacket(whisperTo, text, Opcodes.S_OPCODE_GLOBALCHAT, 9));
+		whisperTo.sendPackets(new S_ChatPacket(whisperFrom, text, Opcodes.S_OPCODE_WHISPERCHAT, 16));
 	}
 
 	@Override
