+++ src/net/l1j/server/model/poison/L1DamagePoison.java
@@ -78,7 +78,7 @@ public class L1DamagePoison extends L1Poison {
 
 	private void doInfection() {
 		_target.setSkillEffect(STATUS_POISON, 30000);
-		_target.setPoisonEffect((byte) 1);
+		_target.setPoisonEffect(1);
 
 		if (isDamageTarget(_target)) {
 			_timer = new NormalPoisonTimer();
@@ -106,7 +106,7 @@ public class L1DamagePoison extends L1Poison {
 			_timer.interrupt(); // 毒タイマー解除
 		}
 
-		_target.setPoisonEffect((byte) 0);
+		_target.setPoisonEffect(0);
 		_target.killSkillEffectTimer(STATUS_POISON);
 		_target.setPoison(null);
 	}
